Sentiment_Analysis-Twitter/Backend/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.4
django-cors-headers==3.11.0
django-rest-swagger==2.2.0
djangorestframework==3.13.1
djongo==1.3.6
idna==3.3
itypes==1.2.0
Jinja2==3.1.2
joblib==1.1.0
MarkupSafe==2.1.1
nltk==3.7
oauthlib==3.2.0
openapi-codec==1.3.2
pymongo==3.12.3
pytz==2022.1
regex==2022.4.24
requests==2.27.1
requests-oauthlib==1.3.1
simplejson==3.17.6
sqlparse==0.2.4
tqdm==4.64.0
tweepy==4.8.0
uritemplate==4.1.1
urllib3==1.26.9
